API testing except for friends.py

+++ rentals.py
@@ -1,13 +1,15 @@
 from db import *
 from methods import *
-from flask import Blueprint, jsonify, request
-from datetime import datetime, timedelta
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask import *
+from datetime import *
+from flask_jwt_extended import *
+from sqlalchemy import *
 
 rentals_bp = Blueprint('rentals', __name__)
 
+
 # 대여 조회(참여 가능한 일정들만)
-@rentals_bp.route('/list', methods=['GET'])
+@rentals_bp.route('/list', methods=['POST'])
 @jwt_required()
 def rentals_list():
     methods_update_rentals()
@@ -20,7 +22,7 @@ def rentals_list():
     if not spaceid or not date_str:
         return jsonify({'error': 'Space ID and date are required'}), 400
 
-    if not (0 <= spaceid or spaceid <= 2) :
+    if not (1 <= spaceid and spaceid <= 3) :
         return jsonify({'error': 'Space ID is invalid'}), 400
 
     try:
@@ -32,12 +34,12 @@ def rentals_list():
     end_of_day = datetime.combine(date, datetime.max.time())
 
     rentals = db.session.query(Rentals).filter(
-        Rentals.spaceid == spaceid,
-        Rentals.starttime >= start_of_day,
-        Rentals.endtime <= end_of_day,
-        Rentals.rentalstatus != Rentals_Status_enum.Close,
-        _or(Rentals.status == Rentals_Status_enum.Light,
-        Rentals.status == Rentals_Status_enum.Club)
+    Rentals.spaceid == spaceid,
+    Rentals.starttime >= start_of_day,
+    Rentals.endtime <= end_of_day,
+    Rentals.rentalstatus != Rentals_Status_enum.Close,
+    or_(Rentals.rentaltype == Rentals_Types_enum.Light,
+    Rentals.rentaltype == Rentals_Types_enum.Club)
     ).all()
 
     rentals_list = []
@@ -69,7 +71,7 @@ def rentals_list():
 
 
 # 대여 참여
-@rentals_bp.route('/join', methods=['GET'])
+@rentals_bp.route('/join', methods=['POST'])
 @jwt_required()
 def rentals_join():
     methods_update_rentals()
@@ -79,33 +81,33 @@ def rentals_join():
     rentalid = data.get('rentalid')
 
     if not rentalid :
-        return jsonify({'error': 'Rental ID are required'}), 400
+        return jsonify({'error': 'Rental ID are required'}), 401
 
     rental = db.session.query(Rentals).filter(
-        Rentals.rentalid == rentalid,
-        Rentals.status != Rentals_Status_enum.Close,
+    Rentals.rentalid == rentalid,
+    Rentals.rentalstatus != Rentals_Status_enum.Close,
     ).first()
 
     if not rental :
-        return jsonify({'error': 'Rental ID is invalid'}), 400
+        return jsonify({'error': 'Rental is not exist'}), 402
 
     # Light 일정에 참여하는 경우
     if rental.rentaltype == Rentals_Types_enum.Light :
-        rentalparticipant = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid, participantid == current_userid).first()
+        rentalparticipant = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid, RentalParticipants.participantid == current_userid).first()
     
         if rentalparticipant : 
-            return jsonify({'error': 'You already participated in'}), 400
+            return jsonify({'error': 'You already participated in'}), 403
 
         rentalparticipant = RentalParticipants(rentalid = rentalid, participantid = current_userid)
         db.session.add(rentalparticipant)
-        rental.poeple += 1
+        rental.people += 1
 
     # Club 일정에 참여하는 경우
     elif rental.rentaltype == Rentals_Types_enum.Club :
-        rentalparticipant = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid, participantid == current_userid).first()
+        rentalparticipant = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid, RentalParticipants.participantid == current_userid).first()
     
         if rentalparticipant : 
-            return jsonify({'error': 'You already participated in'}), 400
+            return jsonify({'error': 'You already participated in'}), 404
 
         # 동아리 인원만 모집중인 경우
         if rental.rentalstatus == Rentals_Status_enum.Half :
@@ -114,16 +116,17 @@ def rentals_join():
             if clubmember : 
                 rentalparticipant = RentalParticipants(rentalid = rentalid, participantid = current_userid)
                 db.session.add(rentalparticipant)
-                rental.poeple += 1
+                rental.people += 1
             
             # 동아리 회원 아닌 경우
             else :
-                return jsonify({'error': 'You cannot join the Club'}), 400
+                return jsonify({'error': 'You cannot join the Club'}), 405
+        
         # 동아리 인원 모집 후 추가 모집
         elif rental.rentalstatus == Rentals_Status_enum.Open :
             rentalparticipant = RentalParticipants(rentalid = rentalid, participantid = current_userid)
             db.session.add(rentalparticipant)
-            rental.poeple += 1
+            rental.people += 1
 
     db.session.commit()
 
@@ -131,31 +134,35 @@ def rentals_join():
 
 
 # 대여 참여 취소
-@rentals_bp.route('/cancle', methods=['GET'])
+@rentals_bp.route('/cancle', methods=['POST'])
 @jwt_required()
 def rentals_cancle():
     methods_update_rentals()
     current_userid = get_jwt_identity()
     data = request.json
 
-    data.get('rentalid')
+    rentalid = data.get('rentalid')
 
     rental = db.session.query(Rentals).filter(Rentals.rentalid == rentalid).first()
     if not rental :
-        return jsonify({'error': 'rental not exist'}), 400
+        return jsonify({'error': 'rental not exist'}), 401
     if rental.rentalflag == Rentals_Flags_enum.Fix :
-        return jsonify({'error': 'rental is fixed'}), 400
-    rental.people -= 1
+        return jsonify({'error': 'rental is fixed'}), 402
+    if rental.userid == current_userid :
+       return jsonify({'error': 'host cannot cancle'}), 404 
     rentalparticipant = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid, RentalParticipants.participantid == current_userid).first()
     if not rentalparticipant :
-        return jsonify({'error': 'rentalparticipant not exist'}), 400
+        return jsonify({'error': 'rentalparticipant not exist'}), 405
     db.session.delete(rentalparticipant)
+    rental.people -= 1
+
     db.session.commit()
 
     return jsonify({}), 200
 
+
 # Light 대여 생성
-@rentals_bp.route('/create', methods=['GET'])
+@rentals_bp.route('/create', methods=['POST'])
 @jwt_required()
 def rentals_create():
     methods_update_rentals()
@@ -172,56 +179,62 @@ def rentals_create():
     friends = data.get('friends',[])
 
     # spaceid 검사
-    if not (0 <= sapceid <= 2) :
+    if not (1 <= spaceid <= 3) :
         return jsonify({'error': 'Invalid spaceid'}), 400
     
     # time 검사
-    if startime >= endtime :
-        return jsonify({'error': 'Invalid starttime and endtime'}), 400
+    if starttime >= endtime :
+        return jsonify({'error': 'Invalid starttime and endtime'}), 401
 
     if not (6 <= starttime.hour <= 22) or not (6 <= endtime.hour <= 22) :
-        return jsonify({'error': 'Invalid starttime and endtime'}), 400
+        return jsonify({'error': 'Invalid starttime and endtime'}), 402
 
     rental = db.session.query(Rentals).filter(
-            Rentals.spaceid == spaceid,
-            _or((Rentals.starttime >= start_datetime,
-            Rentals.starttime <= end_datetime),
-            (Rentals.endtime >= start_datetime,
-            Rentals.endtime <= end_datetime))
+    Rentals.spaceid == spaceid,
+    or_(
+        and_(Rentals.starttime >= starttime, Rentals.starttime <= endtime),
+        and_(Rentals.endtime >= starttime, Rentals.endtime <= endtime)
+    )
     ).first()
     if rental :
-        return jsonify({'error': 'Rental already exist'}), 400
-    
-    # maxpeople 검사
-    sportspace = db.session.query(SportSpace).filter(
-        SportSpace.spaceid == spaceid
-    ).first() 
-    if not (sportspace.minpeople <= maxpeople <= sportspace.maxpeople) :
-        return jsonify({'error': 'Invalid maxpeople value'}), 400
+        return jsonify({'error': 'Rental already exist'}), 403
 
-    # friends 검사 및 초대 알림 생성
-    for friend in friends :
-        user = db.session.query(Users).filter(Users.studentid == friend).first() 
-        if not user :
-            continue
-        friendid = user.userid
-        notification = Notifications(userid = friendid, msg = "Invited to schedule", timestamp = datetime.utcnow(), status = Notifications_ReadStatus_enum.Unread, rentalid = rental.rentalid)
-        db.session.add(notification)
+    # maxpeople 검사
+    sportsspace = db.session.query(SportsSpace).filter(
+        SportsSpace.spaceid == spaceid
+    ).first() 
+    
+    if not (sportsspace.minpeople <= maxpeople <= sportsspace.maxpeople) :
+        return jsonify({'error': 'Invalid maxpeople value'}), 404
 
     # rental 생성
-    rental = Rentals(spaceid = spaceid, userid = current_userid, starttime = starttime, endtime = endtime, createtime = datetime.utcnow(), maxpeople = maxpeople, minpeolple = sportspace.minpeople, people = 1, rentaltype = Rentals_Types_enum.Light, rentalstatus = Rentals_Status_enum.Open, rentalflag = Rentals_Flags_enum.Nonfix, desc = desc)
+    rental = Rentals(spaceid = spaceid, userid = current_userid, starttime = starttime, endtime = endtime, createtime = datetime.now(), maxpeople = maxpeople, minpeople = sportsspace.minpeople, people = 0, rentaltype = Rentals_Types_enum.Light, rentalstatus = Rentals_Status_enum.Open, rentalflag = Rentals_Flags_enum.Nonfix, desc = desc)
     db.session.add(rental)
+    db.session.commit()
 
     # user 를 일정에 추가
     rentalparticipant = RentalParticipants(rentalid = rental.rentalid, participantid = current_userid)
     db.session.add(rentalparticipant)
+    db.session.commit()
+    rental.people += 1
+
+    # 친구들도 일정에 추가
+    for friend in friends :
+        user = db.session.query(Users).filter(Users.studentid == friend).first() 
+        if not user :
+            continue
+        rentalparticipant = RentalParticipants(rentalid = rental.rentalid, participantid = user.userid)
+        db.session.add(rentalparticipant)
+        db.session.commit()
+        rental.people += 1
     
     db.session.commit()
+
     return jsonify({}), 200
 
 
 # 대여 삭제
-@rentals_bp.route('/delete', methods=['GET'])
+@rentals_bp.route('/delete', methods=['POST'])
 @jwt_required()
 def rentals_delete():
     methods_update_rentals()
@@ -235,16 +248,18 @@ def rentals_delete():
     if not rental :
         return jsonify({'error' : 'rental not exist'}), 400
     if rental.rentalflag == Rentals_Flags_enum.Fix :
-        return jsonify({'error' : 'rental is fixed'}), 400    
-    rentalparticipants = db.session.query(Rentals).filter(Rentals.rentalid == rentalid).all()
+        return jsonify({'error' : 'rental is fixed'}), 401    
+    rentalparticipants = db.session.query(RentalParticipants).filter(RentalParticipants.rentalid == rentalid).all()
     for rentalparticipant in rentalparticipants :
         notify_users.append(rentalparticipant.participantid)
         db.session.delete(rentalparticipant)
+        db.sessoion.commit()
     
     # 알림 #
     for notify_user in notify_users :
-        notify = Notifications(userid=notify_user, notifytype=1, spaceid=rental.spaceid, starttime=rental.starttime, endtime=rental.endtime)
+        notify = Notifications(userid=notify_user, notifytype=Notifications_Types_enum.rental_cancle, spaceid=rental.spaceid, starttime=rental.starttime, endtime=rental.endtime)
         db.session.add(notify)
+        db.session.commit()
     # 알림 #
 
     db.session.delete(rental)
